FIx for cache hit test failing on codebetter CI.
git-svn-id: https://flickrnet.svn.codeplex.com/svn@106266 81e6e737-5a20-4d0f-a97c-cb01297b46e0

+++ FlickrNetTest/CacheTests.cs
@@ -36,7 +36,10 @@ namespace FlickrNetTest
         [Test]
         public void CacheHitTest()
         {
-            Directory.Delete(Flickr.CacheLocation, true);
+            if (Directory.Exists(Flickr.CacheLocation))
+            {
+                Directory.Delete(Flickr.CacheLocation, true);
+            }
 
             Flickr f = TestData.GetInstance();
             Flickr.FlushCache();
